:recycle: Split amplitude knob from normalize module
