+++ CHANGELOG.md
@@ -6,3 +6,16 @@ see the commit logs. The format is based on
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+### Fixed
+
+- Returning terminal screen when initialization throws an error.
+
+## 0.0.1 - 2021-05-29
+
+### Added
+
+- Audio file reader and writers.
+- File system navigator.
+- Interactive audio signal plotting.
+- Tab menu for view selection.
